refactor: PayoutSummary 필드 Long타입으로 변경

+++ settlement-service/src/main/java/com/moneykk/moneytown/settlement/query/dto/SettlementBatchDetailResponse.java
@@ -38,10 +38,10 @@ public record SettlementBatchDetailResponse(
     }
 
     public record PayoutSummary(
-            int totalCount,
-            int paidCount,
-            int failedCount,
-            int pendingCount
+            long totalCount,
+            long paidCount,
+            long failedCount,
+            long pendingCount
     ) {
     }
 }
